% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/secure.R
\name{add_cred}
\alias{add_cred}
\title{add credentials}
\usage{
add_cred(server, un, pw, key)
}
\arguments{
\item{server}{name from vmodbc pacakge}

\item{un}{User name of credential}

\item{pw}{password of credential}

\item{key}{encryption key}
}
\value{
TRUE on completion
}
\description{
add credentials
}
