% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/secure.R
\name{get_cred_location}
\alias{get_cred_location}
\title{gets the file name containing credentials location}
\usage{
get_cred_location()
}
\value{
credential file location
}
\description{
gets the file name containing credentials location
}
