% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/secure.R
\name{get_server_cred}
\alias{get_server_cred}
\title{get credentials for a specific server}
\usage{
get_server_cred(server_name, key)
}
\arguments{
\item{server_name}{name of the server in the credentials file}

\item{key}{encryption key}
}
\value{
list with un, pw
}
\description{
get credentials for a specific server
}
