% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/secure.R
\name{make_key}
\alias{make_key}
\title{make an encryption key}
\usage{
make_key(output_file)
}
\arguments{
\item{output_file}{file in which to save key}
}
\value{
encryption key
}
\description{
make an encryption key
}
