% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/secure.R
\name{read_cred}
\alias{read_cred}
\title{read credentials from pacakge data}
\usage{
read_cred(key)
}
\arguments{
\item{key}{encryption key}
}
\value{
credentials
}
\description{
read credentials from pacakge data
}
