% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/secure.R
\name{save_cred}
\alias{save_cred}
\title{encrypt and save credentials using key}
\usage{
save_cred(cred, key)
}
\arguments{
\item{cred}{credentials data.frame}

\item{key}{encryption key}
}
\value{
TRUE on completion
}
\description{
encrypt and save credentials using key
}
