% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/secure.R
\name{setup_cred_file}
\alias{setup_cred_file}
\title{setup empty cred file at file_name}
\usage{
setup_cred_file(file_name, key)
}
\arguments{
\item{file_name}{to save encrypted credentials}

\item{key}{encryption key to use}
}
\value{
file_name
}
\description{
setup empty cred file at file_name
}
