% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/secure.R
\name{update_cred_location}
\alias{update_cred_location}
\title{stores the file_name internally in the package}
\usage{
update_cred_location(file_name)
}
\arguments{
\item{file_name}{cred file name}
}
\value{
file_name
}
\description{
stores the file_name internally in the package
}
